size_t to int

+++ ImageCluster.h
@@ -16,10 +16,10 @@ public:
 	~ImageCluster();
 
 	// load image, and read width and height of image
-	int load_image(unsigned char * _image, size_t _width, size_t _height);
+	int load_image(unsigned char * _image, int _width, int _height);
 
 	// set kernel, and slice the image.
-	int init_kernel_size(size_t kernel_width, size_t kernel_height);
+	int init_kernel_size(int kernel_width, int kernel_height);
 
 	// run the cluster task
 	int cluster();
@@ -30,13 +30,13 @@ private:
 	void check_kernel_vec();
 
 	// check one kernel by x and y
-	bool check_one_kernel(size_t x, size_t y);
+	bool check_one_kernel(int x, int y);
 
 	// get point in image using index in kernel_vec
-	void get_start_point_by_index(size_t _index, size_t &_x, size_t &_y);
+	void get_start_point_by_index(int _index, int &_x, int &_y);
 
 	// get the neighbor points index by index, and the neighbor size is always 8.
-	void get_nearby_points_by_index(size_t _index, std::vector<size_t>& _near_points);
+	void get_nearby_points_by_index(int _index, std::vector<int>& _near_points);
 
 	// delete pointer safely
 	void delete_pointer(unsigned char ** _ptr);
@@ -44,17 +44,17 @@ private:
 private:
 	unsigned char * m_image_data;
 
-	size_t m_width;
+	int m_width;
 
-	size_t m_height;
+	int m_height;
 
-	size_t m_kernel_width;
+	int m_kernel_width;
 
-	size_t m_kernel_height;
+	int m_kernel_height;
 
-	size_t m_kernel_count_col;
+	int m_kernel_count_col;
 
-	size_t m_kernel_count_row;
+	int m_kernel_count_row;
 
 	std::vector<bool> m_kernel_vec;
